+++ ApiConfig.php
@@ -12,14 +12,42 @@ class ApiConfig
      *
      * @var string
      */
-    const VERSION = '0.1.0';
+    const VERSION = '0.5.0';
+
+    /**
+     * API authentication identifier.
+     *
+     * @var string
+     */
+    const DEFAULT_AUTH_IDENTIFIER = 'Api-Key';
+
+    /**
+     * API default hostname.
+     *
+     * @var string
+     */
+    const DEFAULT_HOST_NAME = 'https://api-v3.mobian.global/api';
+
+    /**
+     * API default language.
+     *
+     * @var string
+     */
+    const DEFAULT_LANGUAGE = 'en';
 
     /**
      * Host where the MOBIAN API is located.
      *
      * @var string
      */
-    private static $hostname = 'https://api-v3.mobian.global/api';
+    private static $hostname = self::DEFAULT_HOST_NAME;
+
+    /**
+     * Authentication identifier.
+     *
+     * @var string
+     */
+    private static $authIdentifier = self::DEFAULT_AUTH_IDENTIFIER;
 
     /**
      * Authentication key.
@@ -33,7 +61,7 @@ class ApiConfig
      *
      * @var string
      */
-    private static $language = 'en';
+    private static $language = self::DEFAULT_LANGUAGE;
 
     /**
      * Set the hostname for future requests.
@@ -56,6 +84,26 @@ class ApiConfig
     }
 
     /**
+     * Set the auth identifier for future requests.
+     *
+     * @param string $authIdentifier
+     */
+    public static function setAuthIdentifier($authIdentifier)
+    {
+        self::$authIdentifier = $authIdentifier;
+    }
+
+    /**
+     * Get the current auth identifier.
+     *
+     * @return string
+     */
+    public static function getAuthIdentifier()
+    {
+        return self::$authIdentifier;
+    }
+
+    /**
      * Set the auth key for future requests.
      *
      * @param string $authKey
@@ -68,7 +116,7 @@ class ApiConfig
     /**
      * Get the current auth key.
      *
-     * @return string $authKey
+     * @return string
      */
     public static function getAuthKey()
     {
@@ -93,5 +141,17 @@ class ApiConfig
     public static function getLanguage()
     {
         return self::$language;
+    }
+
+    /**
+     * Reset the current configuration.
+     */
+    public static function reset()
+    {
+        self::$authKey = null;
+
+        self::$authIdentifier = self::DEFAULT_AUTH_IDENTIFIER;
+        self::$hostname = self::DEFAULT_HOST_NAME;
+        self::$language = self::DEFAULT_LANGUAGE;
     }
 }