+++ Requests/Properties/FetchAll.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Mobian\ResellerApi\Requests\Properties;
+
+use Mobian\ResellerApi\Requests\AbstractRequest;
+
+class FetchAll extends AbstractRequest
+{
+    /**
+     * @inheritdoc
+     */
+    protected $method = self::METHOD_GET;
+
+    /**
+     * @inheritdoc
+     */
+    protected $endpoint = '/properties';
+
+    /**
+     * Constructor.
+     *
+     * @param array $query
+     */
+    public function __construct(array $query = [])
+    {
+        // TODO: Implement query parsing
+        $this->params = $query;
+    }
+}