+++ Responses/PlainTextResponse.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Mobian\ResellerApi\Responses;
+
+class PlainTextResponse extends AbstractResponse
+{
+}